+++ vscode/src/custom-prompts/CommandsController.ts
@@ -5,7 +5,6 @@ import { VsCodeCommandsController } from '@sourcegraph/cody-shared/src/editor'
 
 import { logDebug, logError } from '../log'
 import { localStorage } from '../services/LocalStorageProvider'
-import { TreeViewProvider } from '../services/TreeViewProvider'
 
 import { CommandRunner } from './CommandRunner'
 import { CustomPromptsStore } from './CustomPromptsStore'
@@ -44,8 +43,6 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
 
     public commandRunners = new Map<string, CommandRunner>()
 
-    private treeView = new TreeViewProvider('command')
-
     constructor(context: vscode.ExtensionContext) {
         this.tools = new ToolsProvider(context)
         const user = this.tools.getUserInfo()
@@ -56,8 +53,6 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
         this.lastUsedCommands = new Set(localStorage.getLastUsedCommands())
         this.custom.activate()
         this.fileWatcherInit()
-
-        vscode.window.registerTreeDataProvider('cody.commands.tree.view', this.treeView)
     }
 
     /**
