+++ vscode/src/services/TreeViewProvider.ts
@@ -1,12 +1,7 @@
 import * as vscode from 'vscode'
 
-import { UserLocalHistory } from '@sourcegraph/cody-shared/src/chat/transcript/messages'
+import { CodySidebarTreeItem, CodyTreeItemType, getCodyTreeItems } from './treeViewItems'
 
-import { CODY_DOC_URL, CODY_FEEDBACK_URL, DISCORD_URL } from '../chat/protocol'
-
-import { envInit } from './LocalAppDetector'
-
-type CodyTreeItemType = 'command' | 'support' | 'search' | 'chat'
 export class TreeViewProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
     private treeNodes: vscode.TreeItem[] = []
     private _disposables: vscode.Disposable[] = []
@@ -14,7 +9,7 @@ export class TreeViewProvider implements vscode.TreeDataProvider<vscode.TreeItem
     public readonly onDidChangeTreeData = this._onDidChangeTreeData.event
 
     constructor(private type: CodyTreeItemType) {
-        this.updateTree(this.getCodyTreeItems(type))
+        this.updateTree(getCodyTreeItems(type))
         this.refresh()
     }
 
@@ -30,20 +25,6 @@ export class TreeViewProvider implements vscode.TreeDataProvider<vscode.TreeItem
         }
         // TODO implement getParent method for non-root items
         return undefined
-    }
-
-    /**
-     * Gets the tree view items to display based on the provided type.
-     */
-    private getCodyTreeItems(type: CodyTreeItemType): CodySidebarTreeItem[] {
-        switch (type) {
-            case 'command':
-                return commandsItems
-            case 'support':
-                return supportItems
-            default:
-                return []
-        }
     }
 
     /**
@@ -115,115 +96,4 @@ export class TreeViewProvider implements vscode.TreeDataProvider<vscode.TreeItem
         }
         this._disposables = []
     }
-}
-
-interface CodySidebarTreeItem {
-    title: string
-    icon: string
-    id?: string
-    description?: string
-    command: {
-        command: string
-        args?: string[] | { [key: string]: string }[]
-    }
-    isNestedItem?: string
-}
-
-const supportItems: CodySidebarTreeItem[] = [
-    {
-        title: 'Settings',
-        icon: 'settings-gear',
-        command: { command: 'cody.status-bar.interacted' },
-    },
-    {
-        title: 'Keyboard Shortcuts',
-        icon: 'keyboard',
-        command: { command: 'workbench.action.openGlobalKeybindings', args: ['@ext:sourcegraph.cody-ai'] },
-    },
-    {
-        title: 'Release Notes',
-        description: `v${envInit.extensionVersion}`,
-        icon: 'github',
-        command: {
-            command: 'vscode.open',
-            args: [`https://github.com/sourcegraph/cody/releases/tag/vscode-v${envInit.extensionVersion}`],
-        },
-    },
-    {
-        title: 'Documentation',
-        icon: 'book',
-        command: { command: 'vscode.open', args: [CODY_DOC_URL.href] },
-    },
-    {
-        title: 'Feedback',
-        icon: 'feedback',
-        command: { command: 'vscode.open', args: [CODY_FEEDBACK_URL.href] },
-    },
-    {
-        title: 'Discord',
-        icon: 'organization',
-        command: { command: 'vscode.open', args: [DISCORD_URL.href] },
-    },
-    {
-        title: 'Sign Out',
-        icon: 'log-out',
-        command: { command: 'cody.auth.signout' },
-    },
-]
-
-const commandsItems: CodySidebarTreeItem[] = [
-    {
-        title: 'Document',
-        icon: 'book',
-        description: 'Add code documentation',
-        command: { command: 'cody.command.document-code' },
-    },
-    {
-        title: 'Edit',
-        icon: 'wand',
-        command: { command: 'cody.command.edit-code' },
-        description: 'Edit Code with Instructions',
-    },
-    {
-        title: 'Explain',
-        icon: 'output',
-        command: { command: 'cody.command.explain-code' },
-        description: 'Explain code',
-    },
-    {
-        title: 'Smell',
-        icon: 'symbol-keyword',
-        command: { command: 'cody.command.smell-code' },
-        description: 'Identify code smells',
-    },
-    {
-        title: 'Test',
-        icon: 'package',
-        command: { command: 'cody.command.generate-tests' },
-        description: 'Generate unit tests',
-    },
-    {
-        title: 'Custom',
-        icon: 'tools',
-        command: { command: 'cody.action.commands.custom.menu' },
-        description: 'Custom commands',
-    },
-]
-
-// functon to create chat tree items from user chat history
-export function createCodyChatTreeItems(userHistory: UserLocalHistory): CodySidebarTreeItem[] {
-    const chatTreeItems: CodySidebarTreeItem[] = []
-    const chatHistoryEntries = [...Object.entries(userHistory.chat)]
-    chatHistoryEntries.forEach(([id, entry]) => {
-        const lastHumanMessage = entry?.interactions?.findLast(interaction => interaction?.humanMessage)
-        if (lastHumanMessage?.humanMessage.displayText && lastHumanMessage?.humanMessage.text) {
-            chatTreeItems.push({
-                id,
-                title: lastHumanMessage.humanMessage.displayText.split('\n')[0],
-                icon: 'comment-discussion',
-                command: { command: 'cody.chat.panel.restore', args: [id] },
-            })
-        }
-    })
-    return chatTreeItems.reverse()
 }
